requirements.txt
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
autolab-core==1.1.1
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
colorlog==6.10.1
contourpy==1.3.3
cryptography==46.0.3
curl_cffi==0.13.0
cycler==0.12.1
dill==0.4.0
filelock==3.20.0
fonttools==4.60.1
freetype-py==2.5.1
frozendict==2.4.6
fsspec==2025.10.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
ibm-cloud-sdk-core==3.24.2
ibm-platform-services==0.71.0
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy_loader==0.4
lxml==6.0.2
manifold3d==3.2.1
mapbox_earcut==1.0.3
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
multiprocess==0.70.18
multitasking==0.0.12
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
patsy==1.0.2
peewee==3.18.3
pillow==12.0.0
platformdirs==4.5.0
protobuf==6.33.0
psutil==7.1.3
pycollada==0.9.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pyglet==2.1.11
PyJWT==2.10.1
PyOpenGL==3.1.0
pyparsing==3.2.5
pyrender==0.1.45
pyspnego==0.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
qiskit==2.2.3
qiskit-aer==0.17.2
qiskit-algorithms==0.4.0
qiskit-ibm-runtime==0.43.1
referencing==0.37.0
requests==2.32.5
requests_ntlm==1.3.0
rpds-py==0.28.0
rtree==1.4.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
rustworkx==0.17.1
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
setproctitle==1.3.7
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
statsmodels==0.14.5
stevedore==5.5.0
svg.path==7.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.10.16
torch==2.9.0
trimesh==4.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vhacdx==0.0.9
visualization==1.0.0
websockets==15.0.1
xxhash==3.6.0
yfinance==0.2.66
